+++ src/game/Object/PlayerSpell.cpp
@@ -882,179 +882,6 @@ void Player::removeSpell(uint32 spell_id, bool disabled, bool learn_low_rank, bo
 }
 
 /**
- * @brief Removes a cooldown entry for a specific spell.
- *
- * @param spell_id The spell identifier whose cooldown should be removed.
- * @param update True to notify the client that the cooldown was cleared.
- */
-void Player::RemoveSpellCooldown(uint32 spell_id, bool update /* = false */)
-{
-    m_spellCooldowns.erase(spell_id);
-
-    if (update)
-    {
-        SendClearCooldown(spell_id, this);
-    }
-}
-
-/**
- * @brief Removes cooldowns for all spells in a cooldown category.
- *
- * @param cat The spell category identifier.
- * @param update True to notify the client for cleared cooldowns.
- */
-void Player::RemoveSpellCategoryCooldown(uint32 cat, bool update /* = false */)
-{
-    SpellCategoryStore::const_iterator ct = sSpellCategoryStore.find(cat);
-    if (ct == sSpellCategoryStore.end())
-    {
-        return;
-    }
-
-    const SpellCategorySet& ct_set = ct->second;
-    for (SpellCooldowns::const_iterator i = m_spellCooldowns.begin(); i != m_spellCooldowns.end();)
-    {
-        if (ct_set.find(i->first) != ct_set.end())
-        {
-            RemoveSpellCooldown((i++)->first, update);
-        }
-        else
-        {
-            ++i;
-        }
-    }
-}
-
-void Player::RemoveArenaSpellCooldowns()
-{
-    // remove cooldowns on spells that has < 15 min CD
-    SpellCooldowns::iterator itr, next;
-    // iterate spell cooldowns
-    for (itr = m_spellCooldowns.begin(); itr != m_spellCooldowns.end(); itr = next)
-    {
-        next = itr;
-        ++next;
-        SpellEntry const* entry = sSpellStore.LookupEntry(itr->first);
-        // check if spellentry is present and if the cooldown is less than 15 mins
-        if (entry &&
-            entry->GetRecoveryTime() <= 15 * MINUTE * IN_MILLISECONDS &&
-            entry->GetCategoryRecoveryTime() <= 15 * MINUTE * IN_MILLISECONDS )
-        {
-            // remove & notify
-            RemoveSpellCooldown(itr->first, true);
-        }
-    }
-}
-
-/**
- * @brief Clears all tracked spell cooldowns for the player.
- */
-void Player::RemoveAllSpellCooldown()
-{
-    if (!m_spellCooldowns.empty())
-    {
-        ObjectGuid guid = GetObjectGuid();
-
-        WorldPacket data(SMSG_CLEAR_COOLDOWNS, 1 + 8 + m_spellCooldowns.size() * 4);
-        data.WriteGuidMask<1, 3, 6>(guid);
-        data.WriteBits(m_spellCooldowns.size(), 24);      // cooldown count
-        data.WriteGuidMask<7, 5, 2, 4, 0>(guid);
-
-        data.WriteGuidBytes<7, 2, 4, 5, 1, 3>(guid);
-
-        for (SpellCooldowns::const_iterator itr = m_spellCooldowns.begin(); itr != m_spellCooldowns.end(); ++itr)
-        {
-            data << uint32(itr->first);
-        }
-
-        data.WriteGuidBytes<0, 6>(guid);
-
-        SendDirectMessage(&data);
-
-        m_spellCooldowns.clear();
-    }
-}
-
-/**
- * @brief Loads persisted spell cooldowns from the database query result.
- *
- * @param result The query result containing cooldown rows.
- */
-void Player::_LoadSpellCooldowns(QueryResult* result)
-{
-    // some cooldowns can be already set at aura loading...
-
-    // QueryResult *result = CharacterDatabase.PQuery("SELECT `spell`,`item`,`time` FROM `character_spell_cooldown` WHERE `guid` = '%u'",GetGUIDLow());
-
-    if (result)
-    {
-        time_t curTime = time(NULL);
-
-        do
-        {
-            Field* fields = result->Fetch();
-
-            uint32 spell_id = fields[0].GetUInt32();
-            uint32 item_id  = fields[1].GetUInt32();
-            time_t db_time  = (time_t)fields[2].GetUInt64();
-
-            if (!sSpellStore.LookupEntry(spell_id))
-            {
-                sLog.outError("Player %u has unknown spell %u in `character_spell_cooldown`, skipping.", GetGUIDLow(), spell_id);
-                continue;
-            }
-
-            // skip outdated cooldown
-            if (db_time <= curTime)
-            {
-                continue;
-            }
-
-            AddSpellCooldown(spell_id, item_id, db_time);
-
-            DEBUG_LOG("Player (GUID: %u) spell %u, item %u cooldown loaded (%u secs).", GetGUIDLow(), spell_id, item_id, uint32(db_time - curTime));
-        }
-        while (result->NextRow());
-
-        delete result;
-    }
-}
-
-/**
- * @brief Saves active spell cooldowns to the database.
- */
-void Player::_SaveSpellCooldowns()
-{
-    static SqlStatementID deleteSpellCooldown ;
-    static SqlStatementID insertSpellCooldown ;
-
-    SqlStatement stmt = CharacterDatabase.CreateStatement(deleteSpellCooldown, "DELETE FROM `character_spell_cooldown` WHERE `guid` = ?");
-    stmt.PExecute(GetGUIDLow());
-
-    time_t curTime = time(NULL);
-    time_t infTime = curTime + infinityCooldownDelayCheck;
-
-    // remove outdated and save active
-    for (SpellCooldowns::iterator itr = m_spellCooldowns.begin(); itr != m_spellCooldowns.end();)
-    {
-        if (itr->second.end <= curTime)
-        {
-            m_spellCooldowns.erase(itr++);
-        }
-        else if (itr->second.end <= infTime)                // not save locked cooldowns, it will be reset or set at reload
-        {
-            stmt = CharacterDatabase.CreateStatement(insertSpellCooldown, "INSERT INTO `character_spell_cooldown` (`guid`,`spell`,`item`,`time`) VALUES( ?, ?, ?, ?)");
-            stmt.PExecute(GetGUIDLow(), itr->first, itr->second.itemid, uint64(itr->second.end));
-            ++itr;
-        }
-        else
-        {
-            ++itr;
-        }
-    }
-}
-
-/**
  * @brief Calculates the current cost to reset the player's talents.
  *
  * @return The reset cost in copper.
